requirements.txt
albumentations==1.3.0
comet_ml==3.32.0
coremltools==6.1
environs==9.5.0
fastapi==0.89.1
ipython==8.9.0
matplotlib==3.6.2
mss==7.0.1
numpy==1.24.1
onnx==1.13.0
onnxruntime==1.13.1
onnxsim==0.4.13
opencv_python==4.7.0.68
openvino==2022.3.0
paddle==1.0.2
pafy==0.5.5
pandas==1.5.2
Pillow==9.4.0
protobuf==4.21.12
psutil==5.9.4
pycocotools==2.0.6
PyYAML==6.0
requests==2.28.1
seaborn==0.12.2
sentry_sdk==1.14.0
setuptools==65.6.3
tensorflow==2.11.0
tensorflowjs==4.2.0
tensorrt==8.5.3.1
tflite_runtime==2.11.0
tflite_support==0.4.3
thop==0.1.1.post2209072238
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
uvicorn==0.20.0
x2paddle==1.4.0
